+++ geowombat/backends/xarray_.py
@@ -172,6 +172,7 @@ def _check_config_globals(filenames, bounds_by, ref_kwargs):
 
 def warp_open(filename,
               band_names=None,
+              nodata=None,
               resampling='nearest',
               dtype=None,
               return_windows=False,
@@ -186,6 +187,7 @@ def warp_open(filename,
     Args:
         filename (str): The file to open.
         band_names (Optional[int, str, or list]): The band names.
+        nodata (Optional[float | int]): A 'no data' value to set. Default is None.
         resampling (Optional[str]): The resampling method.
         dtype (Optional[str]): A data type to force the output to. If not given, the data type is extracted
             from the file.
@@ -202,6 +204,7 @@ def warp_open(filename,
     ref_kwargs = {'bounds': None,
                   'crs': None,
                   'res': None,
+                  'nodata': nodata,
                   'warp_mem_limit': warp_mem_limit,
                   'num_threads': num_threads,
                   'tap': tap,
@@ -270,6 +273,7 @@ def mosaic(filenames,
            bounds_by='reference',
            resampling='nearest',
            band_names=None,
+           nodata=None,
            dtype=None,
            warp_mem_limit=512,
            num_threads=1,
@@ -290,6 +294,7 @@ def mosaic(filenames,
 
         resampling (Optional[str]): The resampling method.
         band_names (Optional[1d array-like]): A list of names to give the band dimension.
+        nodata (Optional[float | int]): A 'no data' value to set. Default is None.
         dtype (Optional[str]): A data type to force the output to. If not given, the data type is extracted
             from the file.
         warp_mem_limit (Optional[int]): The memory limit (in MB) for the ``rasterio.vrt.WarpedVRT`` function.
@@ -306,6 +311,7 @@ def mosaic(filenames,
     ref_kwargs = {'bounds': None,
                   'crs': None,
                   'res': None,
+                  'nodata': nodata,
                   'warp_mem_limit': warp_mem_limit,
                   'num_threads': num_threads,
                   'tac': None}
@@ -328,11 +334,37 @@ def mosaic(filenames,
             with xr.open_rasterio(fn, **kwargs) as dsb:
 
                 if overlap == 'min':
-                    ds = xr_mininum(ds, dsb)
+
+                    if isinstance(nodata, float) or isinstance(nodata, int):
+
+                        ds = xr.where((ds.mean(dim='band') == nodata) & (dsb.mean(dim='band') != nodata), dsb,
+                                      xr.where((ds.mean(dim='band') != nodata) & (dsb.mean(dim='band') == nodata), ds,
+                                               xr_mininum(ds, dsb)))
+
+                    else:
+                        ds = xr_mininum(ds, dsb)
+
                 elif overlap == 'max':
-                    ds = xr_maximum(ds, dsb)
+
+                    if isinstance(nodata, float) or isinstance(nodata, int):
+
+                        ds = xr.where((ds.mean(dim='band') == nodata) & (dsb.mean(dim='band') != nodata), dsb,
+                                      xr.where((ds.mean(dim='band') != nodata) & (dsb.mean(dim='band') == nodata), ds,
+                                               xr_maximum(ds, dsb)))
+
+                    else:
+                        ds = xr_maximum(ds, dsb)
+
                 elif overlap == 'mean':
-                    ds = (ds + dsb) / 2.0
+
+                    if isinstance(nodata, float) or isinstance(nodata, int):
+
+                        ds = xr.where((ds.mean(dim='band') == nodata) & (dsb.mean(dim='band') != nodata), dsb,
+                                      xr.where((ds.mean(dim='band') != nodata) & (dsb.mean(dim='band') == nodata), ds,
+                                               (ds + dsb) / 2.0))
+
+                    else:
+                        ds = (ds + dsb) / 2.0
 
                 # ds = ds.combine_first(dsb)
 
@@ -377,6 +409,7 @@ def concat(filenames,
            resampling='nearest',
            time_names=None,
            band_names=None,
+           nodata=None,
            dtype=None,
            overlap='max',
            warp_mem_limit=512,
@@ -399,6 +432,7 @@ def concat(filenames,
         resampling (Optional[str]): The resampling method.
         time_names (Optional[1d array-like]): A list of names to give the time dimension.
         band_names (Optional[1d array-like]): A list of names to give the band dimension.
+        nodata (Optional[float | int]): A 'no data' value to set. Default is None.
         dtype (Optional[str]): A data type to force the output to. If not given, the data type is extracted
             from the file.
         overlap (Optional[str]): The keyword that determines how to handle overlapping data.
@@ -418,6 +452,7 @@ def concat(filenames,
     ref_kwargs = {'bounds': None,
                   'crs': None,
                   'res': None,
+                  'nodata': nodata,
                   'warp_mem_limit': warp_mem_limit,
                   'num_threads': num_threads,
                   'tap': tap,
@@ -454,6 +489,7 @@ def concat(filenames,
                                               bounds_by=bounds_by,
                                               resampling=resampling,
                                               band_names=band_names,
+                                              nodata=nodata,
                                               warp_mem_limit=warp_mem_limit,
                                               num_threads=num_threads,
                                               **kwargs))
@@ -468,6 +504,7 @@ def concat(filenames,
                 concat_list.append(warp_open(filenames[tidx],
                                              resampling=resampling,
                                              band_names=band_names,
+                                             nodata=nodata,
                                              warp_mem_limit=warp_mem_limit,
                                              num_threads=num_threads,
                                              **kwargs))