+++ geowombat/radiometry/sr.py
@@ -232,7 +232,9 @@ class LinearAdjustments(object):
                  sensor=None,
                  to='l8',
                  band_names=None,
-                 scale_factor=1):
+                 scale_factor=1,
+                 src_nodata=0,
+                 dst_nodata=0):
 
         """
         Applies a bandpass adjustment by applying a linear function to surface reflectance values
@@ -243,6 +245,8 @@ class LinearAdjustments(object):
             to (Optional[str]): The sensor to adjust to.
             band_names (Optional[list]): The bands to adjust. If not given, all bands are adjusted.
             scale_factor (Optional[float]): A scale factor to apply to the input data.
+            src_nodata (Optional[int or float]): The input 'no data' value.
+            dst_nodata (Optional[int or float]): The output 'no data' value.
 
         Reference:
 
@@ -272,6 +276,9 @@ class LinearAdjustments(object):
         """
 
         attrs = data.attrs.copy()
+
+        # Set 'no data' as nans
+        data = data.where(data != src_nodata)
 
         if scale_factor == 1.0:
             scale_factor = data.gw.scale_factor
@@ -306,6 +313,8 @@ class LinearAdjustments(object):
         if scale_factor != 1:
             data = data / scale_factor
 
+        data = data.fillna(dst_nodata)
+
         data.attrs['adjustment'] = '{} to {}'.format(sensor, to)
         data.attrs['alphas'] = alphas.data.tolist()
         data.attrs['betas'] = betas.data.tolist()
@@ -327,7 +336,8 @@ class RadTransforms(MetaData):
                  solar_az,
                  sensor_za,
                  sensor_az,
-                 nodata=-32768,
+                 src_nodata=-32768,
+                 dst_nodata=-32768,
                  sensor=None,
                  method='srem',
                  angle_factor=0.01,
@@ -342,7 +352,8 @@ class RadTransforms(MetaData):
             solar_az (DataArray): The solar azimuth angle.
             sensor_za (DataArray): The sensor, or view, zenith angle.
             sensor_az (DataArray): The sensor, or view, azimuth angle.
-            nodata (Optional[int or float]): The 'no data' value from the pixel angle data.
+            src_nodata (Optional[int or float]): The input 'no data' value.
+            dst_nodata (Optional[int or float]): The output 'no data' value.
             sensor (Optional[str]): The data's sensor.
             method (Optional[str]): The method to use. Only 'srem' is supported.
             angle_factor (Optional[float]): The scale factor for angles.
@@ -414,12 +425,11 @@ class RadTransforms(MetaData):
                                   sensor_za,
                                   sensor_az,
                                   sensor,
-                                  nodata=nodata)
-
-        sr_data = sr_data.where(sr_data != nodata)
+                                  src_nodata=src_nodata,
+                                  dst_nodata=dst_nodata)
 
         attrs['sensor'] = sensor
-        attrs['nodata'] = nodata
+        attrs['nodata'] = dst_nodata
         attrs['calibration'] = 'surface reflectance'
         attrs['method'] = method
         attrs['drange'] = (0, 1)
@@ -514,7 +524,8 @@ class RadTransforms(MetaData):
                    sensor_za,
                    sensor_az,
                    sensor,
-                   nodata=-32768,
+                   src_nodata=-32768,
+                   dst_nodata=-32768,
                    method='srem'):
 
         """
@@ -527,7 +538,8 @@ class RadTransforms(MetaData):
             sensor_za (DataArray): The sensor zenith angle.
             sensor_az (DataArray): The sensor azimuth angle.
             sensor (str): The satellite sensor.
-            nodata (Optional[int or float]): The 'no data' value from the pixel angle data.
+            src_nodata (Optional[int or float]): The input 'no data' value.
+            dst_nodata (Optional[int or float]): The output 'no data' value.
             method (Optional[str]): The method to use. Currently, only 'srem' is supported and ``method`` is not used.
 
                 Choices:
@@ -545,6 +557,9 @@ class RadTransforms(MetaData):
         """
 
         attrs = toar.attrs.copy()
+
+        # Set 'no data' as nans
+        toar = toar.where(toar != src_nodata)
 
         central_um = toar.gw.central_um[sensor]
         band_names = list(toar.gw.wavelengths[sensor]._fields)
@@ -618,11 +633,20 @@ class RadTransforms(MetaData):
         # Atmospheric backscattering ratio
         ab_ratio = s_atm(r)
 
-        sr_data = (toar_diff / (toar_diff * ab_ratio + transmission)).fillna(nodata).clip(0, 1).astype('float64')
+        sr_data = (toar_diff / (toar_diff * ab_ratio + transmission)).fillna(src_nodata)
+
+        # Create a 'no data' mask
+        mask = sr_data.where((sr_data != src_nodata) & (sr_data > 0)).count(dim='band').astype('uint8')
+
+        # Mask 'no data' values
+        sr_data = xr.where(mask < sr_data.gw.nbands,
+                           dst_nodata,
+                           sr_data.clip(0, 1))\
+            .transpose('band', 'y', 'x').astype('float64')
 
         attrs['sensor'] = sensor
         attrs['calibration'] = 'surface reflectance'
-        attrs['nodata'] = nodata
+        attrs['nodata'] = dst_nodata
         attrs['drange'] = (0, 1)
 
         sr_data.attrs = attrs