+++ geowombat/util/web.py
@@ -27,11 +27,11 @@ from shapely.geometry import Polygon
 try:
 
     from s2cloudless import S2PixelCloudDetector
+
     S2CLOUDLESS_INSTALLED = True
 
 except:
     S2CLOUDLESS_INSTALLED = False
-
 
 shapely.speedups.enable()
 
@@ -41,7 +41,6 @@ RESAMPLING_DICT = dict(bilinear=gdal.GRA_Bilinear,
 
 
 def _rmdir(pathdir):
-
     for child in pathdir.iterdir():
 
         if child.is_file():
@@ -58,7 +57,6 @@ def _rmdir(pathdir):
 
 
 def _assign_attrs(data, attrs, bands_out):
-
     if bands_out:
         data = data.sel(band=bands_out)
 
@@ -69,7 +67,6 @@ def _assign_attrs(data, attrs, bands_out):
 
 
 def _random_id(string_length):
-
     """
     Generates a random string of letters and digits
     """
@@ -80,7 +77,6 @@ def _random_id(string_length):
 
 
 def _parse_google_filename(filename, landsat_parts, sentinel_parts, public_url):
-
     FileInfo = namedtuple('FileInfo', 'url url_file meta angles')
 
     file_info = FileInfo(url=None, url_file=None, meta=None, angles=None)
@@ -90,7 +86,6 @@ def _parse_google_filename(filename, landsat_parts, sentinel_parts, public_url):
     fn_parts = f_base.split('_')
 
     if fn_parts[0].lower() in landsat_parts:
-
         # Collection 1
         url_ = '{PUBLIC}-landsat/{SENSOR}/01/{PATH}/{ROW}/{FDIR}'.format(PUBLIC=public_url,
                                                                          SENSOR=fn_parts[0],
@@ -187,6 +182,7 @@ class GeoDownloads(object):
                       mask_qa=False,
                       lqa_mask_items=None,
                       chunks=512,
+                      cloud_heights=None,
                       num_threads=1,
                       **kwargs):
 
@@ -221,6 +217,7 @@ class GeoDownloads(object):
             mask_qa (Optional[bool]): Whether to mask data with the QA file.
             lqa_mask_items (Optional[list]): A list of QA mask items for Landsat.
             chunks (Optional[int]): The chunk size to read at.
+            cloud_heights (Optional[list]): The cloud heights, in kilometers.
             num_threads (Optional[int]): The number of GDAL warp threads.
             kwargs (Optional[dict]): Keyword arguments passed to ``to_raster``.
 
@@ -258,7 +255,6 @@ class GeoDownloads(object):
         #               overwrite=False)
 
         if not lqa_mask_items:
-
             lqa_mask_items = ['fill',
                               'saturated',
                               'cloudconf',
@@ -267,6 +263,8 @@ class GeoDownloads(object):
 
         angle_kwargs = kwargs.copy()
         angle_kwargs['nodata'] = -32768
+
+        nodataval = kwargs['nodata'] if 'nodata' in kwargs else 65535
 
         angle_infos = dict()
 
@@ -289,17 +287,15 @@ class GeoDownloads(object):
         status = Path(outdir).joinpath('status.txt')
 
         if not status.is_file():
-
             with open(status.as_posix(), mode='w') as tx:
                 pass
 
         # Get bounds from geometry
         if isinstance(bounds, tuple) or isinstance(bounds, list):
-
-            bounds = Polygon([(bounds[0], bounds[3]),   # upper left
-                              (bounds[2], bounds[3]),   # upper right
-                              (bounds[2], bounds[1]),   # lower right
-                              (bounds[0], bounds[1]),   # lower left
+            bounds = Polygon([(bounds[0], bounds[3]),  # upper left
+                              (bounds[2], bounds[3]),  # upper right
+                              (bounds[2], bounds[1]),  # lower right
+                              (bounds[0], bounds[1]),  # lower left
                               (bounds[0], bounds[3])])  # upper left
 
             bounds = gpd.GeoDataFrame([0],
@@ -309,7 +305,6 @@ class GeoDownloads(object):
         bounds_object = bounds.geometry.values[0]
 
         if not out_bounds:
-
             # Project the bounds
             out_bounds = bounds.to_crs(crs).bounds.values[0].tolist()
 
@@ -373,12 +368,12 @@ class GeoDownloads(object):
             year_months[dt1.year] = month_range
         else:
 
-            for y in range(dt1.year, dt2.year+1):
+            for y in range(dt1.year, dt2.year + 1):
 
                 if y == dt1.year:
                     year_months[y] = list(range(dt1.month, 13))
                 elif y == dt2.year:
-                    year_months[y] = list(range(1, dt2.month+1))
+                    year_months[y] = list(range(1, dt2.month + 1))
                 else:
                     year_months[y] = months
 
@@ -441,7 +436,8 @@ class GeoDownloads(object):
                         # Download data
                         if sensor.lower() in ['s2', 's2a', 's2b', 's2c']:
 
-                            load_bands = ['B{:02d}'.format(band_associations[bd]) if bd != 'rededge' else 'B{:01d}A'.format(band_associations[bd]) for bd in bands]
+                            load_bands = ['B{:02d}'.format(band_associations[bd]) if bd != 'rededge'
+                                          else 'B{:01d}A'.format(band_associations[bd]) for bd in bands]
 
                             search_wildcards = ['MTD_TL.xml'] + [bd + '.jp2' for bd in load_bands]
 
@@ -512,9 +508,11 @@ class GeoDownloads(object):
                                 lines = tx.readlines()
 
                             if sensor in ['s2', 's2a', 's2b', 's2c']:
-                                outdir_angles = main_path.joinpath('angles_{}'.format(Path(finfo_dict['meta'].name).name.replace('_MTD_TL.xml', '')))
+                                outdir_angles = main_path.joinpath(
+                                    'angles_{}'.format(Path(finfo_dict['meta'].name).name.replace('_MTD_TL.xml', '')))
                             else:
-                                outdir_angles = main_path.joinpath('angles_{}'.format(Path(finfo_dict['meta'].name).name.replace('_MTL.txt', '')))
+                                outdir_angles = main_path.joinpath(
+                                    'angles_{}'.format(Path(finfo_dict['meta'].name).name.replace('_MTL.txt', '')))
 
                             outdir_angles.mkdir(parents=True, exist_ok=True)
 
@@ -664,7 +662,9 @@ class GeoDownloads(object):
                                                     data_cloudless = data.sel(band=['coastal', 'blue', 'red', 'nir1', 'nir', 'rededge', 'water', 'cirrus', 'swir1', 'swir2'])
 
                                                     # Scale from 0-10000 to 0-1 and reshape
-                                                    X = (data_cloudless * 0.0001).clip(0, 1).data.compute(num_workers=num_threads).transpose(1, 2, 0)[np.newaxis, :, :, :]
+                                                    X = (data_cloudless * 0.0001).clip(0, 1).data\
+                                                            .compute(num_workers=num_threads)\
+                                                            .transpose(1, 2, 0)[np.newaxis, :, :, :]
 
                                                     # Predict clouds
                                                     # Potential classes? Currently, only clear and clouds are returned.
@@ -673,10 +673,14 @@ class GeoDownloads(object):
                                                                              cloud_detector.get_cloud_masks(X),
                                                                              ['mask'])
 
+                                                else:
+
                                                     if bands_out:
+
+                                                        # If there are extra bands, remove them because they
+                                                        # are not supported in the BRDF kernels.
                                                         data = _assign_attrs(data, attrs, bands_out)
 
-                                                else:
                                                     logger.warning('  S2Cloudless is not installed, so skipping Sentinel cloud masking.')
 
                                         if sensor.lower() in ['s2', 's2a', 's2b', 's2c']:
@@ -688,7 +692,8 @@ class GeoDownloads(object):
                                             # Convert TOAR to surface reflectance
                                             sr = rt.toar_to_sr(toar_scaled,
                                                                sza, saa, vza, vaa,
-                                                               rad_sensor)
+                                                               rad_sensor,
+                                                               dst_nodata=nodataval)
 
                                         else:
 
@@ -696,7 +701,9 @@ class GeoDownloads(object):
                                             sr = rt.dn_to_sr(data,
                                                              sza, saa, vza, vaa,
                                                              sensor=rad_sensor,
-                                                             meta=meta)
+                                                             meta=meta,
+                                                             src_nodata=nodataval,
+                                                             dst_nodata=nodataval)
 
                                         # BRDF normalization
                                         sr_brdf = br.norm_brdf(sr,
@@ -704,7 +711,8 @@ class GeoDownloads(object):
                                                                sensor=rad_sensor,
                                                                wavelengths=data.band.values.tolist(),
                                                                out_range=10000.0,
-                                                               nodata=kwargs['nodata'] if 'nodata' in kwargs else 65535)
+                                                               src_nodata=nodataval,
+                                                               dst_nodata=nodataval)
 
                                         if bandpass_sensor.lower() in ['l5', 'l7', 's2', 's2a', 's2b', 's2c']:
 
@@ -712,7 +720,9 @@ class GeoDownloads(object):
                                             sr_brdf = la.bandpass(sr_brdf,
                                                                   bandpass_sensor.lower(),
                                                                   to='l8',
-                                                                  scale_factor=0.0001)
+                                                                  scale_factor=0.0001,
+                                                                  src_nodata=nodataval,
+                                                                  dst_nodata=nodataval)
 
                                         if mask_qa:
 
@@ -720,24 +730,26 @@ class GeoDownloads(object):
 
                                                 if S2CLOUDLESS_INSTALLED:
 
+                                                    wavel_sub = sr_brdf.gw.set_nodata(nodataval,
+                                                                                      nodataval,
+                                                                                      (0, 1),
+                                                                                      'float64')
+
                                                     # Estimate the cloud shadows
-                                                    mask = estimate_cloud_shadows((sr_brdf.sel(band=['nir', 'swir1']) * 0.0001).clip(0, 1).astype('float64'),
+                                                    mask = estimate_cloud_shadows(wavel_sub,
                                                                                   mask,
                                                                                   sza,
                                                                                   saa,
                                                                                   vza,
                                                                                   vaa,
+                                                                                  heights=cloud_heights,
                                                                                   num_workers=num_threads)
 
-                                                    sr_brdf = xr.where(mask.sel(band='mask') == 0,
+                                                    # Update the bands with the mask
+                                                    sr_brdf = xr.where((mask.sel(band='mask') == 0) &
+                                                                       (sr_brdf != nodataval),
                                                                        sr_brdf.clip(0, 10000),
-                                                                       kwargs['nodata'] if 'nodata' in kwargs else 65535).astype('uint16')
-
-                                                else:
-
-                                                    sr_brdf = xr.where(sr_brdf != 0,
-                                                                       sr_brdf.clip(0, 10000),
-                                                                       kwargs['nodata'] if 'nodata' in kwargs else 65535).astype('uint16')
+                                                                       nodataval).astype('uint16')
 
                                                 sr_brdf = _assign_attrs(sr_brdf, attrs, bands_out)
                                                 sr_brdf.gw.to_raster(out_brdf, **kwargs)
@@ -760,19 +772,24 @@ class GeoDownloads(object):
                                                     # Mask non-clear pixels
                                                     sr_brdf = xr.where(mask.sel(band='mask') < 2,
                                                                        sr_brdf.clip(0, 10000),
-                                                                       kwargs['nodata'] if 'nodata' in kwargs else 65535).astype('uint16')
+                                                                       nodataval).astype('uint16')
 
                                                     sr_brdf = _assign_attrs(sr_brdf, attrs, bands_out)
                                                     sr_brdf.gw.to_raster(out_brdf, **kwargs)
 
                                         else:
 
-                                            sr_brdf = sr_brdf.clip(0, 10000).astype('uint16')
+                                            # Set 'no data' values
+                                            sr_brdf = sr_brdf.gw.set_nodata(nodataval,
+                                                                            nodataval,
+                                                                            (0, 10000),
+                                                                            'uint16')
+
                                             sr_brdf = _assign_attrs(sr_brdf, attrs, bands_out)
                                             sr_brdf.gw.to_raster(out_brdf, **kwargs)
 
                                         if write_angle_files:
-
+                                            
                                             angle_stack = xr.concat((sza, saa), dim='band').astype('int16')
                                             angle_stack.attrs = sza.attrs.copy()
                                             angle_stack.gw.to_raster(out_angles, **angle_kwargs)
@@ -932,7 +949,8 @@ class GeoDownloads(object):
         if not search_dict:
 
             if not self.search_dict:
-                logger.exception('  A keyword search dictionary must be provided, either from `self.list_gcp` or the `search_dict` argument.')
+                logger.exception(
+                    '  A keyword search dictionary must be provided, either from `self.list_gcp` or the `search_dict` argument.')
             else:
                 search_dict = self.search_dict
 
@@ -1023,7 +1041,6 @@ class GeoDownloads(object):
                         lines = tx.readlines()
 
                     if Path(down_file).parent.joinpath(fbase + '_MTL.txt').as_posix() + '\n' in lines:
-
                         null_items.append(fbase)
                         continue_download = False